Per Session Connections
Depends on my changes to the main api repo for it to work. I did it by listening for a couple of headers -- either the [ user, pass, token ] tuple, or [ accessToken ]. We then construct a connection based on this, overriding (as it were) the connection formed by this.connection. So fetchSchema etc can still work, and we can require authentication for each request.

We could also disable this on a per-model basis, I suppose...

There is one "TODO" here -- once we login successfully, we know the accessToken. We need a way to get this back to the user (presumably through a response header) but I'm not sure I have access to the response.